Fix error when switching workspace in fullscreen
If the user is watching a video in fullscreen mode and switches
into another workspace, there will be a change in the desktop
size which will be notified to DING because the top bar and,
optionally, any other element that occupies size in the screen
(like dash-to-dock) will be shown now.. But when the user returns
to the workspace where the video is playing in fullscreen,
another size change will be notified. The problem is that the
code detects that the screen is at fullscreen mode and will
"jump over" that monitor, thus not sending the data for that
monitor. If the user has only one monitor, this means that
DING will receive an array of zero elements and will fail.
At this moment, the extension will try over and over again
to relaunch DING, and it will fail over and over again until
the user exits fullscreen mode.

This patch fixes this by sending always the data.