DBusUtils cleanup
This cleanup moves all the interfaces XML text into their own
files, and streamlines the creation of a proxy. This greatly
simplifies adding new proxies in the future.

It also manages the availability of a service, allowing to
detect when a service appears and dissappears, and do things
based on it.