Simplify platformData function
The platformData() function, used to get the parent window and
other data when calling a Nautilus DBus method, is
overengineered. It is designed to work both inside the Shell
code and classic GTK code (while here it is always GTK), and
also allows to receive a parameter with forced values (but in
DING it is always called without parameters).

This MR simplifies the code and removes all these unneeded
things.