dbusUtils: Don't spam about missing tools until they're needed
We're showing notifications about missing tools (such as sushi, that is
not part of most gnome installations) at every startup.

We should instead only notify when the required proxy would be required.

So, only use the proxy property (that triggers the notification) when an
action is performed, without caching the proxy as could be a nice
reminder.