Changed shadow for icon text and support black text
Instead of using four black shadows in the four borders (which
results in a shadow too dark), this patch puts only two in the
upper-left and lower-right corners, and with only a 65% of
opacity. This gives a good legibility with light backgrouns
and keep a good appearance.

Also now it allows to change the text color to black.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/212