Limit the size changes signal
The "size-changed" signal from the desktop manager is emitted
for any window that changes its size. Also, with some themes,
it is also emitted when a window is moved.

The problem with this is that this triggers a lot of dbus
activity in DING, because the extension notifies directly each
signal received.

To avoid this, a time limit has been added, sending up to one
dbus notification per second.