Keep the application if there arent' monitors
If, at any moment, there are no monitors, the number of windows
will be zero and the application will exit. The extension will
relaunch it, of course, but it will produce a CPU overload
until a monitor does appear.

It seems that it is possible to have no monitors in several
cases, like when having an external monitor connected through
an USB-C connector with DisplayPort and disconnecting it.

This patch fixes this problem by adding a hold() call.