Refactor the relaunch code
The relaunch code that process the case when the DING main
process dies had some problems. One was that, in case that
there were a bug in it that prevented it to run, it would try
to relaunch it every 100ms or even faster, which could result
in slow responsiveness of the system or even a hang up.

This patch fixes this, guaranteeing that if the process dies,
it will be relaunched as soon as possible, but also ensuring
that no more than one launch per second is tried.