Remove transient_for
When a popup window is transient for another, both windows will
be moved to the top.

In DING, some error messages are set as transient of the desktop
window, which means that the icons will be shown over other
windows when an error dialog is shown.

This patch removes the use of transient_for to fix this.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/234