promiseUtils: Add latest gjs version of Gio._promisify with further options
Ensuring that in all the versions we target we use the same strategy,
despite the gjs version in the system.

Plus add the ability to promisify without replacing the original method
implementation, but adding a `method_promise` implementation that can be
used without having to change the others that are already here.

This is included in a new file, as it's meant to be shared by both the
extension and ding itself, without having to include more stuff.