Added an installer for ubuntu users
Ubuntu, by default, blocks any DING version different than the
one included in the system. Although this makes sense to avoid
bug reports from people that installed versions not supported by
Canonical, it is an inconvenience to test it.

This MR adds a script that installs DING changing the UUID, so it
would be installed along the official one.

It is mandatory to disable the official one and only then enable
this one.