Fix errors when deleting big folders
When cancelling the decompression of a file with a big number of
files and folders inside (like the linux kernel sources), it
begins to delete the half-decompressed folder and, after some
time, it hangs and there are a lot of apparently unrelated
errors in the log.

This is due to deleting folders asynchronously "in parallel":
the original code is:

    const toDelete = children.map(info =>
        deleteFile(dir.get_child(info.get_name()), info, cancellable, priority));

    await Promise.all(toDelete);

This launches as many "deteFile"s asynchronous calls as files/
folders are in a folder, and since this is recursive, more and
more asynchronous processes are launched. If the number of
files and/or folders is too big, the resources are exhausted
(like file handles) and the system begins to launch hundreds
of errors like

DING: #0   564d4cb29830 i   ding.js:196 (140e7c8cc1f0 @ 927)
DING: (gjs:135908): Gjs-CRITICAL **: Attempting to call back into JSAPI during the sweeping phase of GC. This is most likely caused by>
DING: (gjs:135908): Gjs-CRITICAL **: The offending signal was g-signal on GDBusProxy 0x564d4d1dd0b0.

or

DING: #0   564d4cb29830 i   ding.js:196 (140e7c8cc1f0 @ 927)
DING: (gjs:135908): Gjs-CRITICAL **: Attempting to run a JS callback during garbage collection. This is most likely caused by destroyi>
DING: (gjs:135908): Gjs-CRITICAL **: The offending callback was AsyncReadyCallback().

This patch fixes this bug by serializing the delete process
(although preserving its asynchronous behavior).