desktopGrid: Use the maximized state on startup
During startup (including monitor hotplugging), mutter can get stuck
in an infinite loop trying to adhere to its window constraints logic
for our fixed-size window, if there are two monitors of different
scales in a Wayland session.

When this happens, we don't have `hide_from_window_list` set because
the infinite loop starts before we can even set it. So we need a
solution that works for regular window types.

To solve this we make the window maximized by default. This stops
mutter's constraints logic from wanting to automatically move the
window between monitors.

Yes we should fix mutter, but that will take much longer to get right.

Fixes: https://launchpad.net/bugs/1976204