The last instance of DING will replace any previous one
In some cases, the instance of DING isn't correctly killed when
the extension is disabled; this means that a rogue process
remains, and when the extension is enabled again, that rogue
process won't allow the new one to be launched.

This patch fixes this by allowing any new instance to replace
the old one.