Ensure that the desktop window is detected
In some cases, after returning from a locked screen, the code
can't detect the desktop window, so it remains like a normal
window instead of being managed like the desktop.

This patch starts a three-seconds timer when the desktop process
is launched, and stops it when as many windows as monitors are
detected. But if not enough windows are detected in that time,
it will kill the desktop process, forcing a relaunch.