Add a check for X11Wayland
If the user has configured the system to force GTK to use X11,
the window detection timer will expire because the extension
won't detect the windows, since they are painted through
X11Wayland. This will make the desktop unusable, because it
will reset every three seconds.

This MR adds a check for this and disables the timer if it
detects that it is running on X11Wayland. It also shows a
dialog informing about this, but also allows to not show it
again.