Fix icons position after changing zoom
After changing zoom, the icons and the underlying grid are
misplaced. This is because the desktop window corner is moved to
the corner of the usable area, instead of being kept on the
monitor corner.

This patch fixes this by ensuring that any desktop window is set
again in the right place after a change in the zoom, resolution,
or any other change that could affect it.