Fix "Open in terminal" for Ptyxis
Some terminals, like Ptyxis, not always honor the working
directory as the directory where to open a terminal. To fix
this, this patch also passes the desired directory path using
the --dir= parameter when invoking xdg-terminal-exec.

It also fixes the invocation when passing a command, by removing
the -e parameter in xdg-terminal-exec, since it doesn't use it.