Fix accessibility for .desktop files
After adding the keyboard selection support, the .desktop files
stopped being identified by the true visible name but, instead,
by the file name when using a screen reader.

This patch fixes this by using the file name when the .desktop
file is not trusted (thus, can't be launched), and using the
application name and the "Application" role when it is trusted.

Also, the accessibility code has been refactored and the
sentence generation has been streamlined and reworked to ensure
that it works as expected in any language (and specifically to
avoid problems with right-to-left languages). This,
unfortunately, requires new translations, but in the long term
will be much better.